move two modules into util.py
