+++ src/util.py
@@ -1,6 +1,10 @@
-"""This module returns provides function to return the parameters and types of a function"""
+"""This module returns provides function to return the parameters and types of a
+function and read from return_type and create search strategies that
+compatible with @given()"""
+
 import sqlite3
 import json
+import hypothesis.strategies as st
 
 # TODO: need a function or file to configure the path of database
 # TODO: can it handle types like List[str]?
@@ -56,9 +60,47 @@ def get_input_type(function, module):
         print("Exception in _query: %s" % e)
 
 
-if __name__ == "__main__":
-    types_sorted_output = get_output_type(
-        function="StopWordManager.is_stop_word",
-        module="termfrequency.tf_objectoriented",
-    )
-    print(types_sorted_output)
+def read_type(type_dict):
+    """ Read type from the return types and convert into a list"""
+    if not type_dict:
+        raise Exception("No type given")
+    type_list = list(type_dict.values())
+    return type_list
+
+
+def create_st(type_list):
+    """ Create the compatible arguments for strategies"""
+    st_list = []
+    for t in type_list:
+        if t == "int":
+            st_list.append(st.integers())
+        elif t == "str":
+            st_list.append(st.text())
+        elif t == "float":
+            st_list.append(st.floats())
+        elif t == "bool":
+            st_list.append(st.booleans())
+        # TODO: Enable create_st for complex types
+        # elif t == "object":
+        #     pass
+        # elif t == "List":
+        #     pass
+        # elif t == "Dict":
+        #     pass
+        # elif t == "Tuple":
+        #     pass
+        # elif t == "Union":
+        #     pass
+        # elif t == "Set":
+        #     pass
+        # elif t == "type":
+        #     pass
+    return st_list
+
+
+# if __name__ == "__main__":
+#     types_sorted_output = get_output_type(
+#         function="StopWordManager.is_stop_word",
+#         module="termfrequency.tf_objectoriented",
+#     )
+#     print(types_sorted_output)
